requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3

# Optional dependencies for production
gunicorn==21.2.0
python-dotenv==1.0.0

# For email functionality (optional)
Flask-Mail==0.9.1

# For database (if needed)
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5

# For forms (if needed)
Flask-WTF==1.1.1
WTForms==3.0.1

# For environment variables
python-decouple==3.8